Init project, new scripts
